Make gpg-error-config and gpg-error.m4 more robust.
This is done by taking host prefixed config scripts names in account
(as suggested by bug#1261) and adding a new config option --host to
print a warning if the used config script was not built for the
current host.  Care has been taken not to mess things up if an old
config script without the option is used or if AC_CANONICAL_HOST has
not been used.