Extra test for gpg-error-config script
If the path to gpg-error-config was explicitly given by the user it
may happen that it does not exists but AC_CHECK_TOOL may assume it
still exists because it is somewhere in the PATH.  The extra check
avoids extra sh warnings.

Note that we can't use test -x because that is not defined on some old
systems.