Fix linking for last change on non-ELF platforms.
* src/Makefile.am (libgpg_error_la_LIBADD): Add LIBTHREAD.
--

If configure told us to link with -lpthread we need to pass that
library to the libgpg-error.la build rule.  Which is the right thing
to do.  Without that we may run into strange errors; for example on
AIX pthread_mutex_lock is silently not resolved and thus any use of
gpgrt_lock_lock will immediately lead to segv.

Signed-off-by: Werner Koch <wk@gnupg.org>