Make multi-threading flags available via gpg-error-config.
* m4/threadlib.m4: Set THREADLIB_CPPFLAGS.
* src/gpg-error-config.in: Add option --mt.
* configure.ac: Add support for the --mt option.
* src/gpg-error.m4: Add ac_subst GPG_ERROR_MT_CFLAGS and
GPG_ERROR_MT_LIBS.
--

Although, libgpg-error does not yet provide an API for multi-thread
support, it is useful to add the already available detection to the
config script.  This allows the latest Libgcrypt to take advantage of
this in its regression tests.  In particular for the regression tests
a gpgrt_thread functions would be useful and eventually added to
libgpg-error.  The new gpg-error.m4 script should already be used by
other packages to be prepared for future updates.