Allow ./configure to explicitly set libgpg-error's build timestamp
* configure.ac: add --enable-build-timestamp

--

A group within Debian is working on making the archive rebuildable in
a reproducible way, so that the compiled binary outputs are
byte-for-byte identical when built for the same platform using the
same toolchain.  This is useful in providing auditability and
corroboration for users of the operating system.

libgpg-error is very close to reproducible except for embedding the
build timestamp in the generated binary.  This timestamp is set in
config.h during ./configure.

This patch allows an external build system to set this embedded
timestamp explicitly, which appears to make the package build
repeatably when ./configure is called with (for example)
--enable=build-timestamp=2014-09-23T01:02+0000

Debian-bug-id: 762397

(Minor reformatting and NEWS entry by wk.)