w32: Remove compiler warnings.
* src/Makefile.am (pre_mkheader_cmds): Avoid make diagnostic about
failed but ignored command.  This confuses Emacs' compiler job parser.
* tests/t-lock.c [W32]: Include time.h.
* src/init.c: Reorganize Windows only code.
(wchar_to_utf8, utf8_to_wchar): Remove unused functions.
(_gpg_err_set_errno): Use only one copy for all platforms.
--

Note that there is a still a problem for W64 pertaining to the use of
an int to store a HANDLE.  This will be fixed when we add a full
abstraction layer for Windows file objects.