estream: Avoid calling write(fd,NULL,n).
* src/estream.c (es_func_fd_write): Take care of a flush requests.
(es_func_w32_write): Ditto.
(es_func_fp_write): Ditto.
--

The write handlers may be called with (BUFFER,SIZE) of (NULL,0) to
propagate flush events to user supplied functions (es_fopencookie).
However we need to take care to also do this with the internal
handler.  Might be a reason for system faults due to write or fwrite
with a NULL buffer; on Linux this seems to be harmless.

Signed-off-by: Werner Koch <wk@gnupg.org>