Add function gpgrt_annotate_leaked_object.
* src/gpg-error.h.in: add gpgrt_annotate_leaked_object to support
  marking memory as non-leaked for Clang and GCC.
--
This annotation can be used to mark objects as explicitly leaked such
that it can be ignored in tools like LeakSanitizer.

The GPGRT_HAVE_LEAK_SANITIZER macro is explicitly not undefined to
support -fsanitize=leak, a user or configure script could then decide to
add this macro when just -fsanitize=leak is given.

Signed-off-by: Peter Wu <peter@lekensteyn.nl>

Additional changes by -wk:

  - But __GNUC__ guard around the entire GPGRT_HAVE_LEAK_SANITIZER
    detection.
  - Add NEWS entry.

Signed-off-by: Werner Koch <wk@gnupg.org>