w32: Remove cruft in w32-estream from olden times.
* src/w32-io.c (set_synchronize): Remove.
(create_reader, create_writer): No need for set_synchronize.
--

The set_synchronize dates back to 2001 at a time when gpgme got
the Windows support on WindowsME and Windows2000.  Maybe this was
required then due to bugs in that old NT or partly NT based Windows
versions.

Removal similar to the removal in gpgme's w32-io.