build: Fix installation of gpg-error-config.1.
* doc/Makefile.am (myman_pages): Rename man page to gpg-error-config.1
and only install if gpg-error-config itself is installed.
--

The man page embedded within doc/gpgrt.texi has recently been renamed
from gpgrt-config.1 to gpg-error-config.1, but the Makefile.am was still
referring to the old name, resulting in a `make install` failure.

Also make the man page installation conditional on
--enable-install-gpg-error-config, as there is no point in having the
man page if the corresponding tool is not installed.

Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>