m4: Fix detection of gpgrt's libdir.
* src/gpg-error.m4 (GPGRT_CONFIG): Handle the case where none of the
system lib directories contain a pkgconfig subdirectory.
--

When we look for gpgrt_libdir, there's a corner case if we had been able
to obtain system libdirs (using `cc -print-search-dirs`) *but* none of
those system libdirs happen to contain a valid pkgconfig subdirectory
(which may be unlikely but can and does happen when cross-compiling).

We do test for the case where we have not obtained any system libdir at
all, in which case we fallback to `${gpgrt_prefix}/lib`
(`possible_libdir1`), but we do not test if the list of libdir
candidates is reduced to nothing after we have eliminated all the
libdirs that do not contain a pkgconfig subdirectory.

This patch adds a test for this precise case.

Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>