logging: New function gpgrt_logv_domain.
* src/gpg-error.h.in (gpgrt_logv_domain): New.  Also add printf
attributes to the other logv functions.
* src/visibility.c (gpgrt_logv_domain): New.
--

This function can be used instead of gpgrt_logv to provide a somain
string.  As of now this has no effect but we may later add an API to
allow selection of diagnostics depending on the domain.  There is also
a new (buffer,length) argument which will eventually be used to
provide a domain based gpgrt_log_printhex.