Add source file strlist.c
--

This is a copy from current GnuPG master.  The license has been
chnaged from (LGPL-3.0-or-later OR GPL-2.0-or-later) to
LGPL-2.1-or-later: Almost everything was written by me or by employees
of g10 Code GmbH.  The only exception is strlist_pop

  2002-06-14  David Shaw  <dshaw@jabberwocky.com>

  * strgutil.c (pop_strlist): New function to pop the head off of a
  strlist.

from gnupg-1.4/util/strgutil.c.  However the function is to obvious to
inhibit the license change.  It needs to be changes anyway with
followup commits.

Signed-off-by: Werner Koch <wk@gnupg.org>