w32: Improve the detection of the need for a \\?\ prefix
* src/sysutils.c (_gpgrt_fname_to_wchar): Use the no-max-path prefix
for shorter files and for filenames with path separators.
--

Depending on the current directory the old threshold of 230 was to
larger.  Now we use a lower limit which should handle most standard
file names w/o the need for the prefix.  We also always do this if the
requested file includes a directory part.

Updates-commit: 32dad4e4135f0bdc436ed684da753d4b0cdb0ea1