w32: Fix extended length path handling for UNC specified files.
* src/sysutils.c (_gpgrt_fname_to_wchar): Fix the extended length path
support for UNC file names.  Also add an envvar to disable the entire
extended length path support.
--

GnuPG-bug-id: 5754
Fixes-commit: a4a692fcf1ed128693ec6d64d5e3163775aad10b

With the last fix we fixed the extended length path support but it
also truned outa wrong implementaion of it for UNC.  Before the
previous patch the UNC support was already wrong but it exhibited
itself only when the requested UNC was already long.  The patch then
kicked in the support more often and in particular always if an UNC was
used (due to the new checking for backslashes and not just the
length).

To avoid further surprises we now have an envvar to disable that
support entirely.  But if should be correct now ;-).