Time for a new error code; this time GPG_ERR_UNEXPECTED_PACKET
--

We already have _tag which is used for ASN.1, _MSG which is used by
protocols like WKS and TLS, and the GPG_ERR_UNEXPECTED which is
unrelated the former one and means just that an error was not
expected.  OpenPGP also has some cases where a packet is not expected.
A dedicated error code thus makes sense.